+++ src/main/java/br/com/mercadolivre/bootcampw2/grupo11/socialmeli/services/UsersService.java
@@ -27,7 +27,7 @@ public class UsersService {
     public UserDTO getUserInfo(Integer userId) {
         return userRepository.findById(userId)
                 .map(UserDTO::fromEntity)
-                .orElseThrow(() -> new ResourceNotFoundException("User", userId.longValue()));
+                .orElseThrow(() -> new ResourceNotFoundException("User", userId));
     }
